Add files via upload
At one point there were about 60 or 70 of these, maybe more. But in '99 my sister formatted the computer they were on in the process of updating Windows, causing literally all of my work to be deleted (thousands of hours of logic and math), except for whatever I happened to have saved on 3.5" floppies at the time. Here are the surviving ASCII animations.